requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
pillow